Refactor extract decrypt tally to a dedicated class/file
